requirements.txt
-i https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cpu
absl-py==2.1.0; python_version >= '3.7'
asgiref==3.8.1; python_version >= '3.8'
django==5.1.6; python_version >= '3.10'
djangorestframework==3.15.2; python_version >= '3.8'
filelock==3.17.0; python_version >= '3.9'
fsspec==2025.2.0; python_version >= '3.8'
h5py==3.13.0; python_version >= '3.9'
jinja2==3.1.5; python_version >= '3.7'
keras==3.8.0; python_version >= '3.9'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
mdurl==0.1.2; python_version >= '3.7'
ml-dtypes==0.5.1; python_version >= '3.9'
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2; python_version >= '3.10'
numpy==2.2.3; python_version >= '3.10'
optree==0.14.1; python_version >= '3.8'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
psycopg[binary,pool]==3.2.5; python_version >= '3.8'
psycopg-binary==3.2.5; python_version >= '3.8'
psycopg-pool==3.2.6; python_version >= '3.8'
pygments==2.19.1; python_version >= '3.8'
pyjwt==2.10.1; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2025.1
rich==13.9.4; python_full_version >= '3.8.0'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlparse==0.5.3; python_version >= '3.8'
sympy==1.13.1; python_version >= '3.8'
torch==2.6.0+cpu
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2025.1; python_version >= '2'
